test-print.R: don't assume model path names are under 250 characters
A test assumes that calling print.bbi_process() with its default
.call_limit value of 250 will be sufficient to get at least two models
in the printed call string, but it's possible for that to not be
enough space to avoid those paths being truncated.

Check whether the paths of the first two models exceed 250 and use a
non-default .call_limit if needed.