fix: print.bbi_process: avoid marking untruncated args as truncated
To decide whether to truncate, print.bbi_process() checks whether the
length of the full call string, including executable path and trailing
long options, is greater than .call_limit.  But the truncation is
limited to the positional arguments that follow the executable.  That
means that the positional argument substring may not be larger than
.call_limit, in which case feeding it to substr(..., 1, .call_limit)
won't lead to a truncated result, despite the "... [truncated]" marker
being appended.

Add another condition to guard against this case.