tail_output: give informative message for finished run
When tail_output() is called on a NONMEM model that has already
finished, the downstream read_lines() call in check_file() errors with
an unhelpful message about the file not existing.  Instead check the
model status and give a message when the model has already finished
running.

Note that looking at the status before calling
check_file()/read_lines() would be more readable than using tryCatch,
but that's susceptible to a race with the running model.

Closes #408.