test: add helper to skip test for old bbi versions
If a bbr test depends on a new behavior or feature in bbi (perhaps
unreleased), it should be written in a way that doesn't fail if bbi
points to an older version that is still at least bbr.bbi_min_version.
Add a skip helper that can be used in this case.

The bbr test will often be added before a new bbi release has been
cut, in which case

  * the only bbi version that is known for sure is the current release
    that's incompatible.  This is the reason for making the argument
    the latest _incompatible_ version.

  * the bbi version string before release will have a "develop" prefix
    and can't be compared.  Don't skip in this case because this will
    often be a recent development build that has the given feature.
    Any even if the build is older and incompatible, it's a manual
    run, so the caller can inspect and process the result.