refactor: namespace cleanup
The objective is to remove the library loading during tests. The (minor) concern
is that the loading of the library could potentially pollute the viability of test results
and hide improper missing imports etc in the actual code base. That potentially
is not a problem given R CMD CHECK's static analysis, however
my gut reaction is to feel that everything in test should be externally clear where
its coming from.