feat: param_estimates.bbi_nonmem_summary: account for ONLYSIM
`bbi nonmem summary` used to fail with an out-of-bounds error when fed
ONLYSIM output, but it now sets the "only_sim" field.  With the bbi
failure removed, param_estimates() fails for ONLYSIM models:

    > param_estimates(s)
    Error: All columns in a tibble must be vectors.
    ✖ Column `parameter_names` is NULL.
    ✖ Column `estimate` is NULL.
    ✖ Column `fixed` is NULL.

Make param_estimates.bbi_nonmem_summary() abort if it sees the
only_sim flag, similar to what it already does for Bayesian methods.