use-bbi: prefer built-in R functionality to system commands
To download and install bbi, use_bbi() makes a sequence of system()
calls that depend on the system's wget rm, tar, mv, and chmod
programs.  This has a few drawbacks:

  * The commands rely on external executables that, even in *nix
    environments, might not be available.

  * On Windows a different code path (windows_install_commands) is
    executed than on *nix systems (nx_install_commands), increasing
    the odds that the less frequently exercised Windows code bitrots
    or has unnoticed bugs.

  * Code injection is possible via use_bbi() arguments.  In the
    context of bbr, which isn't exposed to untrusted outside input,
    this is likely not much of a security concern.  However, it's
    still a good idea to avoid it.

Deal with these three issues by replacing the system() calls with a
shared download/install function that uses calls to built-in R
functions.  It's of course still possible to have OS-specific issues
with these functions, but it's less unlikely because 1) they are
widely used and 2) they are designed to be portable (see, e.g., the
Windows-specific notes in the help pages of download.file and untar).