fix: bbi_version: fix handling of executable paths with spaces
The use_bbi() rewrite in the last commit sidestepped a few issues
related to file names with spaces.  However, use_bbi() calls
bbi_version(), and that still trips up on on files names with spaces
because it joins the bbi_executable path and "version" with a
space (and no special quoting) and calls system() with the string,
which will fail if the executable has spaces.

Use processx::run() instead, which doesn't require the command or
arguments be quoted.

Fixes #409.