test: skip_if_old_bbi: treat any invalid version as a dev build
skip_if_old_bbi() has support for local bbi builds by checking for the
"develop-" prefix, but it will fail on some tags we use (e.g.,
"-rc1").  If the version doesn't look a valid version expected for an
official release, assume we're testing a developmental build and don't
skip the test.