requirements: append "works" to bare function name descriptions
The initial requirement descriptions were taken from the associated
test names.  When a test description is rolled up, this often means
that the description is just the test name.  While that may be okay
for test descriptions, especially when a function only has one
associated test ID, having a requirement description that is just a
function name is probably more likely to cause confusion when people
are reviewing the validation docs.

Tack on "works" to match the style that is used for lots of other
functions.