test-read-output.R: rearrange descriptions to differentiate tests
When mrgvalprep rolls up the descriptions in test-read-output.R, the
distinction between some of the different tests are lost.  For
example, BBR-ROT-005 and BBR-ROT-006 come out the same:
"check_output_dir()".

Rework the test descriptions to get different and more informative
rolled-up descriptions.

In the process, adjust the case table so that they include only the
part of the description that is different between the cases.  This
makes it quicker to identify what's different and avoids unintended
differences between the repeated prefix.