docs: regenerate with latest roxygen2
This repo's pkgr.yml pulls from the latest CRAN, so the generated docs
should generally be built with the latest roxygen2 available.
roxygen2 v7.2.1 came out 2022-07-18 and produces a lot off Rd changes,
most of which are probably related the describedIn changes [*].

Regenerate the documentation in a dedicated PR so that we can quickly
merge these changes and allow in-flight PRs to build on top of
them (avoiding conflicts and duplicate commits for the same Rd
changes).

[*] https://github.com/r-lib/roxygen2/blob/main/NEWS.md#roxygen2-721