test-refs/1_summary_obj.R: run through RStudio's "Reformat Code"
I'm not sure how these files were generated.  If I do

  dput(model_summary(read_model(MOD1_PATH)),
       SUMMARY_REF_FILE)

the representation is more compressed and has trailing spaces, so I
suspect there was some reformatting already.  Send it through
RStudio's "Reformat Code" command (in "Code" menu).  This expands the
representation a bit, which is good for producing diffs that highlight
the changes, and hopefully provides a way for others to generate
consistent formatting.