model_summaries.list: pass multiple models to bbi if supported
`bbi nonmem summary` accepted multiple models since v2.3.0.  The
upcoming v3.2.0 release reworks the output so that the errors can be
obtained from the JSON data.

When a new enough bbi is available, route model_summaries() through a
single call to bbi to take advantage of bbi concurrency.  This should
make model_summaries() perform better when operating on a large number
of models (e.g., in a bootstrap setting).  Note that it's possible for
this to exceed the maximum argument length for the suprocess.  If we
run into that in practice, we can consider adding logic to split up
the arguments, or modifying bbi to accept the model paths on stdin.

The intent is to align the serial and concurrent versions closely,
with downstream code being blind to which implementation was used.  I
decided not to port the "may be because the modeling run has not
finished" addition that model_summary() prepends to bbi error
messages, because I'm not sure it's very useful and it's easy enough
to add if desired.

Closes #395.