check_run_times: use test_bbi_version
As of bc19f34f (warning message and handling for bbi versions <=
3.1.1, 2022-07-06), check_run_times() has somewhat unusual treatment
of a bbi version condition in that it warns that the bbi is
incompatible but continues.  This spot followed skip_if_old_bbi() and
used the last compatible version as the reference (at my suggestion).
With the last commit, skip_if_old_bbi() now takes the minimum required
version, and, for the reasons mentioned there, it makes sense for
check_run_times() to do the same.

Update this spot to use test_bbi_version(), going through the same
logic that's triggered by bbi_exec().