test-refs/1_summary_obj.R: regenerate with development bbi
The JSON structure of `bbi nonmem summary` has changed.  With a bbi
built from e1e1bda (v3.2.0-alpha.3, 2022-07-12), call

  dput(model_summary(read_model(MOD1_PATH)),
       SUMMARY_REF_FILE)

and then reformat the file with RStudio's "Reformat Code" command.

This prunes a recently added comment, but that's all right because
these generated files aren't a good place for comments and the issue
it is commenting on is already explained in bbi's d13e0e1 (nmparser:
don't overwrite covariance time with postprocess time, 2022-07-06).