test-summary-log.R: set bbi_exe_path via file-wide
test-summary-log.R sets bbr.bbi_exe_path to the result of
read_bbi_path() via with_options().  BBR-SMLG-011 is outside of this
block.  An upcoming commit will make model_summaries() check the bbi
version, which requires the bbi path being set, leading to
BBR-SMLG-011 failing (depending on the particular environment).

None of the tests in this file depend on modifying the
bbr.bbi_exe_path value, so set the option file-wide to avoid people
needing to worry about whether the a given test needs to be within the
block.

Aside from the 'with_options -> local_options' change, the entire diff
is due to unindenting the with_options() block.