fix: replace check_lst_file() with build_path_from_model(.x, ".lst") in   bbi_exec_model_summaries()
This is needed because otherwise model_summaries() (and its dependants) will error if an .lst file is missing (i.e. if one of the models failed to run) instead of returning the list of models, with the missing .lst error passed through in the correct spot.

Essentially, we want bbi to be the thing that errors, so that we can catch those errors and put them into the returned list.