refactor: get rid of check_lst_file().
This helper is quite old and the motivation for it is no longer relevant. It also complicates the way we return errors in model_summary(), and makes it different than how we do it in model_summaries().

Essentially we want to faithfully pass the command through to bbi and then pass back any errors that result.