model_summaries: handle NULL .fail_flags
Both the original model_summaries() implementation (now named
model_summaries_serial) and the new implementation try to make a
second call even when .failed_flags isn't specified.  Repeating the
call in this case doesn't make sense, though it should just result in
a pointless call (except for if an in-progress run happens to complete
in between the two calls).

However, both implementations have the same combine_list_objects()
calls, and that fails if .failed_flags is NULL, so the second call is
never made.  (For the original implementation, the error actually gets
presented as a model error.)  Make both implementations return early
if .failed_flags isn't specified, avoiding the failure and unnecessary
repeated bbi call.

Note that, with the code reorganization in
model_summaries_concurrent(), the positions in $Errors are relied on
by downstream code, exposing a logic bug: errors positions for the
retried call need to be mapped back to the original positions.