print_bbi_args: use cli to format message
The output of print_bbi_args() is a little difficult to digest because
the descriptions can be fairly long and the items bleed into each
other.

Elsewhere in bbi we use cli to format messages.  Do the same with this
output.  Make the different items stand out by bolding the argument
names and indenting any subsequent lines that go past the console
width.

This also fixes the lack of a trailing newline in a plain R
console (as opposed to RStudio), which led to the last line mangling
the start of the next prompt.