bbi args: support printing compatibility notes
Some of the items in BBI_ARGS were either changed or added in a more
recent bbi version than bbr.bbi_min_version.  Teach print_bbi_args()
to inform the user.

Another option would be to tailor the display of print_bbi_args() to
the installed bbi version, but 1) that's probably not worth the effort
and 2) it seems helpful let users of older versions know what _could_
be available.