copy_model_from: rework to support unknown model types
copy_model_from_impl() expects model to either be "nonmem" or "stan",
but hard coding the model list doesn't work for case where another
packages depends on bbr and adds a new model type.

Expect for the copy_model_from_impl() helper to be shared by the
extension packages, and allow for a model-type specific setup function
to be specified.