check_up_to_date.bbi_log_df: absorb .build_data into '...' parameter
check_up_to_date.bbi_log_df() has an explicit .build_data argument,
but none of the other check_up_to_date() methods in bbr do.  This is
argument is now documented in bbr.bayes , and, aside from an internal
kludge in config_log_entry(), isn't used here.

It might make sense to document the existence of this argument even
though it's not relevant for NONMEM models, but, until we do so, don't
leak it the signature of check_up_to_date.bbi_log_df().