ci: switch one build to using minimum required bbi version
bbr.bbi_min_version is set to 3.0.2, but Drone won't flag an
incompatibility because all of the Drone runs use the latest bbi
version.  Switch the mpn:oldest run over to using the minimum required
version.

Re-purposing the mpn:oldest run loses the ability to catch some
things (e.g., an interaction between an mpn:latest package and the
minimum bbi, or an mpn:oldest package and the latest bbi), but blowing
out the test matrix to catch these interactions seems like an
overkill.