WOP-patch: Add additional logic in `submit_model` to handle when threads is set to NULL.
Bug is raised in example:

```
MOD1 %>% submit_model(.bbi_args = list(parallel = F, threads = NULL))
```

The error raised is :

```
Error in if (.args$threads > 1 && isFALSE(.args$parallel)) warning("`threads > 1` but model will not run in parallel because `parallel = FALSE`") :
missing value where TRUE/FALSE needed .
```

New logic is added `utils.R` to `check_bbi_args` to catch bug. It is
still a work in progress.

The test for this bug is `BBR-UTL-016` and was added to `UTL-R003` in
the requirements.