update-handling-threads:
In `utils.R`
Revert logic using a series of if-else blocks to cover combintations  if
parallel is set or not or  if threads is set or not. It seemed to make
more sense to be more specific on a case-by-case basis. The `overwrite` parameter is being passed to
catch if threads is NULL and to set number of threads to 1 in the
command line arguments.