revisions:
`utils.R`

Add simplified logic to catch arguments passed if parallel flag.

```
  if (!is.null(.args$threads) && .args$threads > 1) {
    if (is.null(.args$parallel)) {
      cmd_args <- c(cmd_args, "--parallel")
    } else if (isFALSE(.args$parallel)) {
      warning("`threads > 1` but model will not run in parallel because `parallel = FALSE`")
    }
  }
```
It is easier to understand that the else-if block used in previous
commits.

`test-submit-models.R`

Reverted tests that tested when `--threads=1`. This was being passed as
a placeholder until handling defaults was more discussed. It appears
that threads is looked for in the bbi_args and then the model's
defition and then in the bbi.yaml

`test-utils.R`
Added simplifed test for `BBR-UTL-016`