nm-file: don't let custom warn value break fread warning handling
When nm_file_impl() calls data.table::read(), it catches any warning
and looks at the message to decide if multiple tables were in the
file, in which case it warns and returns NULL.  That logic fails if
the user has tweaked the warn option to, for example, silence warnings
or to convert warnings to errors.

Override the warn value to 1 so that the handler can reliably process
the warnings.

Note that the default warn value is 0, which says to show a message
like "There were NN warnings (use warnings() to see them)" if there
are over 10 messages.  That setting works fine because each handler
will still receive the individual warnings, but it also has no effect.
So 1 is simpler value that communicates the behavior we're relying on.