build_path_from_model: don't require output directory to exist
build_path_from_model() fails if the output directory doesn't exist.
On the other hand, it does _not_ fail if the path itself doesn't exist
but the output directory does.  So for consistency it should work if
the output directory doesn't exist.  In other words, this failure
looks like an oversight rather a need for build_path_from_model() to
learn a .check_exists=FALSE option.

Note that model_summaries_concurrent() uses build_path_from_model();
this change fixes a model_summaries() error that's triggered when the
bbi version is at least v3.2.0 and one of the models doesn't have an
output directory.  (The model_summaries() tests will be adjusted in
the next commit to check this case.)