add compatibility wrapper for dplyr joins expecting multiple matches
nm_join() and delete_models() have left_join() calls that trigger the
following dplyr 1.1.0 warning:

  Each row in `x` is expected to match at most 1 row in `y`.
  ℹ Row 1 of `x` matches multiple rows.
  ℹ If multiple matches are expected, set `multiple = "all"` to
    silence this warning.

The behavior is the same before and after dplyr 1.1.0: return all
matches.

For delete_models(), it looks multiple matches are expected for both
left_join() calls.

For nm_join(), multiple matches are not expected for most of its
left_join() calls.  The exception is the left_join(TABLE, DATA) call
for the FIRSTONLY case because there tables have only a row per
individual rather than a row for each data record.

To avoid warnings when multiple matches are expected while staying
compatible with dplyr versions before 1.1.0, add a left_join() wrapper
that passes multiple=all when a new enough dplyr is available.

Only use this wrapper is cases where it's clear multiple return values
are expected because the only consequence of missing a spot is users
seeing a warning (which they will hopefully report, so that we can add
a test case to cover it).