param_estimates_compare: use reframe with dplyr 1.1.0 or later
param_estimates_compare() calls summarise(across(...)) to get a row
for each 'probs' value of a quantile() call.  As of dplyr 1.1.0,
returning more than one row with summarise() is deprecated in favor of
the new reframe() function.  Add a compatibility kludge that maps to
reframe() when it is available.

This kludge is only used in one spot, so it'd be okay to define it in
param-estimates-batch.R until needed elsewhere, but instead add a
dedicated compat.R file and put it there because an upcoming commit
will add another dplyr 1.1.0 kludge, and it's nice for them to live in
the same place.  And in general having a dedicated spot for
compatibility code makes it easier to review bits that are no longer
needed when bumping minimum required versions.