use_bbi: abort if specified path doesn't end with .exe on Windows
Sys.which() and processx::run() won't be able to work with bbi binary
if it doesn't have an ".exe" on disk.

Note that passing "bbi" to Sys.which()/processx::run() would work as
long as "bbi.exe" was on disk, so another option would be to accept
"bbi" here and just make sure we download to an "exe" path.  I think
explicitly requiring the ".exe" has the advantage of making it clear
what the file will be on disk.