use_bbi: fix handling of relative paths
If a relative path is passed to use_bbi(), it fails with a dev error:

  Error in dev_error(glue("Must pass an absolute path but got: `{.path}`")) :
    Must pass an absolute path but got: `bin/bbi`

install_menu() presumably expects that path to be absolute at this
point because it passes the user-specified value to
`normalizePath(..., mustWork = FALSE)`.  But that's not a valid
assumption because, if the path does _not_ exist, normalizePath() will
return that path as given when mustWork=FALSE.

The documentation says that path should be absolute, so one option
would be to add a guard to use_bbi() that gives a regular error, not a
dev error.  But I don't see a strong reason to forbid a relative path,
so remove that limitation by using fs::path_abs(), which does not
require that the path exists on disk.

We could also rework this to not expand the path to an absolute path,
which looks like it should work but would increase the chance of an
oversight/regression given the downstream code was written assuming an
absolute path.