use_bbi: tighten 'bbi in PATH?' check
add_to_path_message() checks whether the specified path is within PATH
by splitting Sys.getenv("PATH") and looking for the parent directory.
One problem with that approach is the splitting and checking require
more extensive processing to do right (e.g., do either the specified
path or the PATH elements contain a symlink or otherwise need to be
normalized before comparing?  Do we need to worry about a PATH element
containing an escaped separator?)

Another problem is that, even if PATH contains the parent directory,
it doesn't tell us what we actually want to know, which is whether a
call with just the base name resolves to the given executable path.
Use Sys.which() to check that, avoiding the need to parse and process
PATH on our end.