use_bbi: widen 'bbr.bbi_exe_path configured?' check
add_to_path_message() checks whether the specified path is equal to
the value of the bbr.bbi_exe_path option.  However, this check isn't
sufficient to decide whether bbr.bbi_exe_path points to the given path
because add_to_path_message() is called with a path that's been
processed with normalizePath().

This misfire is easy to hit on *nix systems, but it's especially
likely to be triggered on Windows because normalizePath() transforms
"/" to "\", and it's valid for the user to use "/" for the path
argument and bbr.bbi_exe_path.

Fix the check by feeding both the given path and the bbr.bbi_exe_path
value to fs::path_real() before comparing.  Note that we could use
normalizePath(..., mustWork = TRUE), but fs::path_real() has the
advantage of signaling a specific error when the file doesn't exist.

Re: #576