ci: pin oldest pipeline to nmrec 0.2.0
The previous commit dropped the custom command for installing nmrec
from most pipelines, but it left the oldest pipeline as is because
that pipeline uses the 2020-06-08 snapshot, and nmrec isn't available
until the 2023-09-19 snapshot.

To follow the "test the oldest supported version" spirit of this
pipeline, check out the 0.2.0 tag of nmrec instead of the tip of main.
0.2.0 is the oldest version that bbr should care about being
compatible with because that's the first version on MPN.