test-workflow-bbi.R: let runner override bbi_init values
test-workflow-bbi.R is the one spot that actually runs the model after
calling bbi_init().  For the run, it uses nm74gf under /opt/NONMEM/
and bbi's default mpiexec (/usr/local/mpich3/bin/mpiexec).

These tests target Metworx, so embedding these details is reasonable,
but it's useful to be able to override them.

 * .nonmem_version: at some point Metworx may drop nm74gf.

   Allow the runner to set the value via an environment variable.

 * mpiexec path: a new Metworx may stop putting mpiexec at the
   non-standard location.

   Use the mpiexec on PATH as a more reliable value (on the current
   Metworx this ultimately points to the same spot as
   /usr/local/mpich3/bin/mpiexec).  Also allow the value to be set via
   an environment variable.

 * .nonmem_dir: this value is unlikely to change on Metworx, but allow
   it to be set via an environment variable for consistency.

The main point of allowing the values to be overridden is to make the
upgrade path easier; if the runner can override them, a bbr update
isn't required for the new Metworx.  It also makes testing more
convenient (e.g., to check if the tests still pass with a different
NONMEM, the tester can set BBR_TESTS_NONMEM_VERSION rather than
changing the hard-coded value in the test file).