model-summaries: rename serial and concurrent helpers for clarity
model_summaries_serial() and model_summaries_concurrent() are variants
focused on using `bbi nonmem summary` underneath.  Both have special
handling for retrying on failure with the set of flags specified in
.fail_fails.  And, as of the last commit, they are only used for plain
NONMEM models, which we know are destined for bbi.

Rename these to include "bbi" because to make this clearer.