docs: don't generate Rd for model_summaries_concurrent
In bbr, the docs of unexported functions are marked with the keyword
'internal'.  That keyword stops pkgdown from complaining about a
function's generated docs not being referenced in the index.

The style.tidyverse.org and rOpenSci guidelines recommend marking
internal functions with noRd rather than 'keywords internal',
preventing the generation of Rd and html pages that are unlikely to be
helpful.

While it may not be worth doing a package-wide transition, it makes
sense to move to noRd as docs are updated.
model_summaries_concurrent() is about to be renamed, so switch it over
to noRd.