tail_*: don't hard code target when passed 'model path'
The character methods for tail_output() and tail_lst() take the path
as a model path if it doesn't look like a full target (i.e. an OUTPUT
file or *.lst file) and then construct the full target from the model
path.  That logic, however, couples the methods too closely with a
specific model structure and fails for nmbayes models, which have
multiple versions of these files in a chain submodels.

Read in the model and dispatch to the model method, letting it
implement whatever custom logic may be appropriate.

Re: https://github.com/metrumresearchgroup/bbr.bayes/issues/36