working prototype for creating a new bbr bootstrap model object
 - After some experimentation, I think we may want to revisit how we want to actually submit a bootstrap model for a few reasons. To check the status of a bootstrap model submission, we iterate through all models defined in the bootstrap control stream. This can delay printing of the model object in the R console (especially for 1000+ runs). It would be better to have a new `bbi.json` file that gets generated after a successful bootstrap model submission, so we only have to check for the existance of/read in one file.

In the interest of not wanting to clutter GHE or SVN repos, we also dont want to commit all the individual model files needed to call model summary. I think it would be better to have bbi collect all the summary results and create a new set of files in a single directory, and then cleanup/delete all the extra ones (files per bootstrap run) behind the scenes. Given time constraints this may have to be a future patch, but wanted to document some reasoning here.