remove stale validation setup
For the GitHub Actions switch, we're moving to
r-lib/actions/setup-r@v2, which uses pak underneath.  The pak
dependency resolution fails due to mrgvalidate and mrgvalprep.

The reason for the failure isn't clear to me, but given that this
project no longer uses these packages for validation, just remove them
as dependencies.  Also drop the stale yaml files related to the old
validation setup.

Closes #629.