test batch submission
 - This seems to fail on occasion (rarely) due to carrying through an embedded nul. Actual error:

  > In readLines(proc$get_output_file()) :
  line 16 appears to contain an embedded nul

If this happens, the model `.lst` file becomes corrupted and cannot be viewed directly by many R functions or via the Rstudio file viewer.

Seth and I observed this once before, though the OUTPUT file was still readable in that case. Will look into this further, but it may require a more in-depth investigation/dedicated PR on top of this one.