submit_model.bbi_nmboot_model now checks .overwrite upfront and ignores .bbi_args and bbi.yaml setting overwrite.
The primary motivation for this is practicality: these runs are often quite large, so making the user explicitly pass an overwrite value for them seems safer and less likely to lead to a time-consuming mistaken overwrite. This also simplifies the code paths necessary for checking cascading overwrite values before the method passes off to bbi further down the call stack.

One additional point is that this brings it in line with current bbr.bayes behavior for bbi_nmbayes_model objects (as of bbr.bayes 0.2.0). The bbr.bayes behavior is currently undocumented, which we should correct, but I think the same logic as above supports that being a reasonable behavior there too.