Move overwrite deletion until _after_ dry_run check.
This corrects an error in a previous commit (742e75f364a7a4b250126df744bc9d61bf016f65) obviously we should not delete an existing directory on a "dry run".

This fixes some of the test failures, though as of this commit, I am still seeing some failures locally in the test-workflow-* files.