DESCRIPTION: prune unused imports
The last direct use of stringi was removed in 04e5a4e0 (remove
ascii_boxplot function, 2024-05-30), part of the 0.11.0 release.

The Matrix import was added in c6f3ef0d (list_c was not an exported
function in `purrr` for MPN oldest, 2024-01-31), part of the 1.9.0
release, but it's unclear why given that Matrix is not used in that
commit.