test-submit-model{,s}.R: future-proof "inherits bbr.bbi_exe_mode" tests
The default mode is "sge" on Linux and "local" otherwise, so the test
for inheriting bbr.bbi_exe_mode makes sure to use a non-default value
so that it can detect that the override was effective.

Looking forward, "sge" will eventually be removed, so rework the
condition to instead check "local" and use "slurm" if that's the
default.

While touching these lines, convert the 'switch' (which sprawls across
four lines when formatted with styler) to an arguably simpler 'if'.