submit_model: expose .mode for Slurm
The upcoming bbi 3.4.0 adds a 'bbi nonmem run slurm ...' command.
Register "slurm" as a valid mode.

Keep "sge" as the default (on Linux) for now but add some guards to
help catch cases where "sge" is specified on a machine that doesn't
support SGE.

Slurm includes a qsub shim that has some behavioral differences when
used under 'bbi nonmem run sge ...' (e.g., output files going to the
parent directory).  That shim would very likely cause confusion when
people on Slurm machines forget to override the "sge" default.

Prevent that confusion on Metworx or other systems using
ParallelCluster by aborting if "sge" is specified for .mode and qsub
is Slurm's shim.  That's aggressive and perhaps will be triggered by
people that have already been making of the shim.  However, it's
better to be strict about it because it's not an option we intend to
support.