add_dofv: convert from S3 method to regular function
add_dofv is a confusing name for the .bbi_nonmem_model method that
takes a model and returns a single value (versus _adding_ a column to
the run log).  The model method is probably of limited use, so for now
go with a regular function that takes a run log.

Re: https://github.com/metrumresearchgroup/bbr/pull/731#discussion_r2046860837