add_dofv: rework to limit number of 'bbi nonmem summary' calls
add_dofv calls model_summary() for each model in the run log and for
each reference model.  Each model_summary() call spawns a bbi
subprocess.

Speed that up by

 1) reusing OFVs from the input when the caller passes a summary log

 2) pass the entire all models that are missing OFVs to a
    model_summaries call (via summary_log_impl), leading to a single
    bbi subprocess

In a benchmark with four models [*], this cuts the add_dofv execution
time by over half when a run log is passed.  When a summary log is
passed, it shaves off over 90% of the execution time.

[*] https://github.com/metrumresearchgroup/bbr/pull/731#issuecomment-2809506811