add_dofv tests: tweak OBJV lines for better expected values
The add_dofv tests use the set of models generated by
create_all_models.  Those are all copies of the same model, so the
expected dofv value is always 0 (or NA), making it impossible to
detect that the correct reference model has been used.

Update the tests to tweak the lst files so that each test model has a
different OBJV value.  While at it, extend the tests 1) to check that
the same dofv values are calculated for a plain run log and the
summary log counterpart, 2) to the check the non-NULL .mod_ref case,
and 3) to check a log where some of the parent models have been
filtered out.

Re: https://github.com/metrumresearchgroup/bbr/pull/731#discussion_r2046866211