tests: add helper to create directory with tweaked dummy models
Setting up many "completed" models and tweaking various details (e.g.,
the objective function value or final estimation method) will be
required to thoroughly test the upcoming add_dofv function and, to a
lesser extent, the addition of AIC and BIC to the summary log.

The existing summary_log tests rely on the create_all_models helper.
Setting up the different scenarios with this helper is possible but
unwieldy.

Add a helper that enables declaring the layout of the model tree and
the various tweaks to be made to each model.