summary_log: add AIC and BIC columns
Extend the summary log with columns for AIC and BIC because 1) these
are values users commonly want to inspect and end up calculating in
custom code and 2) information for the calculation is already
available in the model summary output.

The main complication is that we want to disable reporting AIC and BIC
for methods where the values are unlikely to be meaningful.  The set
of methods selected here is based on SME input.  (The same method
selection will be relevant for the add_dofv function introduced in an
upcoming commit.)

Re: #677

Co-authored-by: Kyle Barrett <barrettk@metrumrg.com>
Helped-by: Samuel Callisto <samc@metrumrg.com>
Helped-by: Kyle Baron <kyleb@metrumrg.com>
Helped-by: Tim Waterhouse <timw@metrumrg.com>