log: add and export function to extend log with dOFV
As of a few commits back, AIC and BIC are to the summary log.
Calculating a change in objective function value is more involved
because it requires pairing a row's model with a reference model.

Expose a function that users can call to extend an existing summary
log (or run log) with a column for dOFV.

Reporting dOFV is disabled for the same set of methods for which
reporting AIC and BIC is.  Because dOFV is explicitly comparing two
models, there are additional restrictions (chosen based on SME input):
the final methods must be identical and the data must look the same.

Re: #677

Co-authored-by: Kyle Barrett <barrettk@metrumrg.com>
Helped-by: Samuel Callisto <samc@metrumrg.com>
Helped-by: Kyle Baron <kyleb@metrumrg.com>
Helped-by: Tim Waterhouse <timw@metrumrg.com>